MEDASSIST-feature-FinetuneV1/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pdfplumber
sentence-transformers
chromadb
requests
